Implemented User autentication and overral refactor of the codee

+++ src/main/java/com/will/bookstoreapi/repository/CategoryRepository.java
@@ -1,10 +1,9 @@
 package com.will.bookstoreapi.repository;
 
+import com.will.bookstoreapi.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.will.bookstoreapi.domain.Category;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-
+public interface CategoryRepository extends JpaRepository<Category,Long> {
 }
